feat: enhance offers and orders management with sorting and filtering options
- Added sorting functionality for offers and orders by date and desired completion date.
- Implemented search filtering for offers to include offer numbers.
- Updated UI to include new sorting and filtering controls.
- Enhanced offer details to display desired completion date and offer duration.
- Improved order details to show desired completion date.
- Added customer response status display in offers.
- Updated settings page to include logo upload functionality and offer link validity days.
- Introduced Telegram bot settings for notifications.
- Implemented database wipe functionality with confirmation steps.
- Updated types to accommodate new fields for offers, orders, and notifications.

+++ backend/src/api/routes/products.ts
@@ -13,7 +13,7 @@ const CreateProductBody = z.object({
   heightMm: z.number().int().nonnegative(), // changed allowing 0 for m2_unsorted
   widthMm: z.number().int().nonnegative(), // changed allowing 0 for m2_unsorted
   calcMethod: z.enum(['m2_unsorted', 'm2_sorted', 'volume_divided']).default('m2_sorted'),
-  volumeDivider: z.number().int().positive().optional(),
+  volumeDivider: z.number().positive().optional(),
   description: z.string().max(2000).optional(),
   // Accept both field names
   initialPricePerM2: z.number().positive().optional(),
@@ -28,7 +28,7 @@ const UpdateProductBody = z.object({
   heightMm: z.number().int().nonnegative().optional(),
   widthMm: z.number().int().nonnegative().optional(),
   calcMethod: z.enum(['m2_unsorted', 'm2_sorted', 'volume_divided']).optional(),
-  volumeDivider: z.number().int().positive().nullable().optional(),
+  volumeDivider: z.number().positive().nullable().optional(),
   description: z.string().max(2000).optional(),
   isActive: z.boolean().optional(),
   currentPricePerM2: z.number().positive().optional(),
